Blagasgda

+++ oving3/local_src/driver-gamepad-1.0/driver-gamepad.c
@@ -234,7 +234,7 @@ static ssize_t my_read (struct file *filp, char __user *buffer, size_t length, l
 }
 
 static void button_map(void) {
-    int fuck;
+    uint8_t fuck;
     fuck = ~ioread8(gpio_portc_mem + DIN_OFFSET);
     printk(KERN_DEBUG "GAMEPAD: mapping buttons\n");
     switch(fuck) {
